app/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
idna==2.6
paho-mqtt==1.3.1
psycopg2==2.7.4
redis==2.10.6
requests==2.18.4
rq==0.10.0
urllib3==1.22
